requirements.txt
anyio==4.4.0
asgiref==3.7.2
attrs==23.2.0
CacheControl==0.14.0
cachetools==5.4.0
certifi==2023.7.22
cffi==1.16.0
channels==4.1.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.5
defusedxml==0.8.0rc2
distlib==0.3.7
Django==4.2.7
django-admin-soft-dashboard==1.0.12
django-cors-headers==4.2.0
django-cors-middleware==1.5.0
django-environ==0.11.2
django-extensions==3.2.3
django-filter==23.3
django-flatpickr==2.0.1
django-jazzmin==3.0.0
django-push-notifications==3.1.0
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
django_rest_resetpassword==0.1.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
dnspython==2.6.1
drf-nested-forms==1.1.8
drf-spectacular==0.27.1
drf-yasg==1.21.7
email_validator==2.2.0
fastapi==0.111.1
fastapi-cli==0.0.4
filelock==3.13.1
firebase-admin==6.5.0
GDAL==3.8.3
geographiclib==2.0
geopy==2.4.1
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.17.0
google-cloud-storage==2.18.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
grpcio==1.65.1
grpcio-status==1.62.2
gunicorn==21.2.0
h11==0.14.0
h3==3.7.7
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
msgpack==1.0.8
numpy==2.0.1
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==23.2
pandas==2.2.2
Pillow==10.1.0
pipenv==2023.10.24
platformdirs==3.11.0
proto-plus==1.24.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
pydantic==1.10.13
Pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.1.2
pyscopg2==66.0.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-http-client==3.3.7
python-multipart==0.0.9
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
razorpay==1.4.1
referencing==0.33.0
reportlab==4.2.2
requests==2.31.0
requests-oauthlib==1.3.1
rest-framework-simplejwt==0.0.2
rich==13.7.1
rpds-py==0.17.1
rsa==4.9
sendgrid==6.11.0
setuptools==69.0.3
shellingham==1.5.4
simplejson==3.19.1
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.0
social-auth-core==4.5.0
SQLAlchemy==2.0.31
sqlparse==0.4.4
starkbank-ecdsa==2.2.0
starlette==0.37.2
stripe==6.0.0
timezonefinder==6.5.2
typer==0.12.3
typing_extensions==4.8.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.0.7
uvicorn==0.30.1
uvloop==0.19.0
virtualenv==20.24.6
watchfiles==0.22.0
websockets==12.0
wheel==0.42.0
